Move PerasEnabled wrapper to O.C.Peras.Params
This commit moves the Maybe-like PerasEnabled and PerasEnabledT wrappers
into O.C.Peras.Params. This is in anticipation of the upcoming HFC
plumbing that reuses these types outside of HFC history queries.
