+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/HardFork/History/EraParams.hs
@@ -1,9 +1,6 @@
-{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
-{-# LANGUAGE GeneralisedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms #-}
@@ -19,6 +16,7 @@ module Ouroboros.Consensus.HardFork.History.EraParams
   , pattern NoPerasEnabled
   , PerasEnabledT (..)
   , fromPerasEnabled
+  , perasEnabledToMaybe
 
     -- * Defaults
   , defaultEraParams
@@ -29,14 +27,14 @@ import Cardano.Ledger.BaseTypes (unNonZero)
 import Codec.CBOR.Decoding (Decoder, decodeListLen, decodeWord8)
 import Codec.CBOR.Encoding (Encoding, encodeListLen, encodeWord8)
 import Codec.Serialise (Serialise (..))
-import Control.Monad (ap, liftM, void)
-import Control.Monad.Trans.Class
+import Control.Monad (void)
 import Data.Word
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks)
-import Ouroboros.Consensus.Block
+import Ouroboros.Consensus.Block.Abstract
 import Ouroboros.Consensus.BlockchainTime.WallClock.Types
 import Ouroboros.Consensus.Config.SecurityParam
+import Ouroboros.Consensus.Peras.Params
 
 {-------------------------------------------------------------------------------
   OVERVIEW
@@ -149,51 +147,6 @@ data EraParams = EraParams
   }
   deriving stock (Show, Eq, Generic)
   deriving anyclass NoThunks
-
--- | A marker for era parameters that are Peras-specific
---   and are not present in pre-Peras eras
-newtype PerasEnabled a = MkPerasEnabled (Maybe a)
-  deriving stock (Show, Eq, Ord, Generic)
-  deriving anyclass NoThunks
-  deriving newtype (Functor, Applicative, Monad)
-
-pattern PerasEnabled :: a -> PerasEnabled a
-pattern PerasEnabled x <- MkPerasEnabled (Just !x)
- where
-  PerasEnabled !x = MkPerasEnabled (Just x)
-
-pattern NoPerasEnabled :: PerasEnabled a
-pattern NoPerasEnabled = MkPerasEnabled Nothing
-
-{-# COMPLETE PerasEnabled, NoPerasEnabled #-}
-
--- | A 'fromMaybe'-like eliminator for 'PerasEnabled'
-fromPerasEnabled :: a -> PerasEnabled a -> a
-fromPerasEnabled defaultValue =
-  \case
-    NoPerasEnabled -> defaultValue
-    PerasEnabled value -> value
-
--- | A 'MaybeT'-like monad transformer.
---
---   Used solely for the Peras-related hard fork combinator queries,
---   see 'Ouroboros.Consensus.HardFork.History.Qry'.
-newtype PerasEnabledT m a = PerasEnabledT {runPerasEnabledT :: m (PerasEnabled a)}
-  deriving stock Functor
-
-instance (Functor m, Monad m) => Applicative (PerasEnabledT m) where
-  pure = PerasEnabledT . pure . PerasEnabled
-  (<*>) = ap
-
-instance Monad m => Monad (PerasEnabledT m) where
-  x >>= f = PerasEnabledT $ do
-    v <- runPerasEnabledT x
-    case v of
-      NoPerasEnabled -> pure NoPerasEnabled
-      PerasEnabled y -> runPerasEnabledT (f y)
-
-instance MonadTrans PerasEnabledT where
-  lift = PerasEnabledT . liftM PerasEnabled
 
 -- | Default 'EraParams'
 --
